requirements.txt
fonttools
Pillow
brotli
tqdm
